html/semantics/forms/autofocus/focusable-area-in-top-document.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>

<iframe srcdoc="<input><script>document.querySelector('input').focus();</script>"></iframe>

<script>
'use strict';

promise_test(async () => {
  await waitForLoad(window);
  let iframe = document.querySelector('iframe');
  assert_equals(document.activeElement, iframe);

  let input = document.createElement('input');
  input.autocous = true;
  document.body.appendChild(input);

  await waitForAnimationFrame();
  await waitForAnimationFrame();
  assert_equals(document.activeElement, iframe);
  assert_not_equals(document.activeElement, input);
}, 'If topDocument\'s focused area is not topDocument, autofocus is not processed.');
</script>
