third_party/blink/web_tests/external/wpt/html/semantics/forms/autofocus/focusable-area-in-top-document-expected.txt

This is a testharness.js-based test.
FAIL If topDocument's focused area is not topDocument, autofocus is not processed. assert_equals: activeElement should not be changed expected Element node <iframe srcdoc="<input><script>document.querySelector('in... but got Element node <input autofocus=""></input>
Harness: the test ran to completion.
